Criado controller para o user e criadas exceptions personalizadas

+++ src/main/java/com/projeto/demo/controllers/UserController.java
@@ -1,6 +1,8 @@
 package com.projeto.demo.controllers;
 
 import com.projeto.demo.dto.UserRegisterDto;
+import com.projeto.demo.entities.User;
+import com.projeto.demo.exceptions.UnauthorizedActionException;
 import com.projeto.demo.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -9,19 +11,10 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping("/user")
-public class UserController {
+public class UserController extends BaseController {
 
     @Autowired
     private UserService userService;
-
-    @PostMapping("/")
-    public ResponseEntity<?> create(@RequestBody UserRegisterDto userDto) {
-        try {
-            return ResponseEntity.ok(userService.register(userDto));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-        }
-    }
 
     @PutMapping("/")
     public ResponseEntity<?> update(@RequestBody UserRegisterDto userDto) {
@@ -32,17 +25,8 @@ public class UserController {
         }
     }
 
-    @GetMapping("/{cpf}")
-    public ResponseEntity<?> getEmail(@PathVariable String cpf) {
-        try {
-            return ResponseEntity.ok(userService.findByCpf(cpf));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
-        }
-    }
-
-    @GetMapping
-    public ResponseEntity<?> list() {
+    @GetMapping("/")
+    public ResponseEntity<?> listUsers() {
         try {
             return ResponseEntity.ok(userService.listUsers());
         } catch (Exception e) {
@@ -50,19 +34,31 @@ public class UserController {
         }
     }
 
-    @DeleteMapping("/cpf/{cpf}")
-    public ResponseEntity<Object> deleteUser(@PathVariable String cpf) {
+    @GetMapping("/id/{id}")
+    public ResponseEntity<?> findById(@PathVariable Long id) {
         try {
-            userService.deleteByCpf(cpf);
-            return ResponseEntity.ok().build();
+            return ResponseEntity.ok(userService.findById(id));
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+        }
+    }
+
+    @GetMapping("/cpf/{cpf}")
+    public ResponseEntity<?> findByCpf(@PathVariable String cpf) {
+        try {
+            return ResponseEntity.ok(userService.findByCpf(cpf));
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
         }
     }
 
     @DeleteMapping("/id/{id}")
-    public ResponseEntity<Object> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<Object> deleteById(@PathVariable Long id) {
         try {
+            if (!isAdmin()) {
+                throw new UnauthorizedActionException();
+            }
+
             userService.deleteById(id);
             return ResponseEntity.ok().build();
         } catch (Exception e) {
@@ -70,4 +66,17 @@ public class UserController {
         }
     }
 
+    @DeleteMapping("/cpf/{cpf}")
+    public ResponseEntity<Object> deleteByCpf(@PathVariable String cpf) {
+        try {
+            if (!isAdmin()) {
+                throw new UnauthorizedActionException();
+            }
+
+            userService.deleteByCpf(cpf);
+            return ResponseEntity.ok().build();
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(e.getMessage());
+        }
+    }
 }
